+++ DevTools/StoreOnline-EPAM.postman_collection.json
@@ -1,0 +1,283 @@
+{
+	"info": {
+		"_postman_id": "2df0730f-d55f-4f2e-930f-f8b4dded5b6d",
+		"name": "StoreOnline-EPAM",
+		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
+		"_exporter_id": "2859979"
+	},
+	"item": [
+		{
+			"name": "Category",
+			"item": [
+				{
+					"name": "List",
+					"request": {
+						"method": "GET",
+						"header": [],
+						"url": {
+							"raw": "https://localhost:7167/api/categories?PageNumber=1&PageSize=10",
+							"protocol": "https",
+							"host": [
+								"localhost"
+							],
+							"port": "7167",
+							"path": [
+								"api",
+								"categories"
+							],
+							"query": [
+								{
+									"key": "PageNumber",
+									"value": "1"
+								},
+								{
+									"key": "PageSize",
+									"value": "10"
+								}
+							]
+						}
+					},
+					"response": []
+				},
+				{
+					"name": "Get",
+					"request": {
+						"method": "GET",
+						"header": [],
+						"url": {
+							"raw": "https://localhost:7167/api/categories/8",
+							"protocol": "https",
+							"host": [
+								"localhost"
+							],
+							"port": "7167",
+							"path": [
+								"api",
+								"categories",
+								"8"
+							]
+						}
+					},
+					"response": []
+				},
+				{
+					"name": "Create",
+					"request": {
+						"method": "POST",
+						"header": [],
+						"body": {
+							"mode": "raw",
+							"raw": "{\r\n    \"name\":\"Lorem ipsum dolor sit amet, consectetur porttitor2212121221212\",\r\n    \"image\":\"https://imagenes.noticiasrcn.com/cms/2024/08/21165207/televisores-por-menos-de-quinientos-mil-pesos-en-falabella.webp?r=4_3\",\r\n    \"parentCategoryId\": 8\r\n}",
+							"options": {
+								"raw": {
+									"language": "json"
+								}
+							}
+						},
+						"url": {
+							"raw": "https://localhost:7167/api/categories/9",
+							"protocol": "https",
+							"host": [
+								"localhost"
+							],
+							"port": "7167",
+							"path": [
+								"api",
+								"categories",
+								"9"
+							]
+						}
+					},
+					"response": []
+				},
+				{
+					"name": "Update",
+					"request": {
+						"method": "PUT",
+						"header": [],
+						"body": {
+							"mode": "raw",
+							"raw": "{\r\n    \"id\":8,\r\n    \"name\":\"Ropa\",\r\n    \"image\":\"https://offloadmedia.feverup.com/valenciasecreta.com/wp-content/uploads/2018/03/13121547/segunda-mano-valencia.jpg\",\r\n    \"parentCategoryId\": null\r\n}",
+							"options": {
+								"raw": {
+									"language": "json"
+								}
+							}
+						},
+						"url": {
+							"raw": "https://localhost:7167/api/categories/8",
+							"protocol": "https",
+							"host": [
+								"localhost"
+							],
+							"port": "7167",
+							"path": [
+								"api",
+								"categories",
+								"8"
+							]
+						}
+					},
+					"response": []
+				},
+				{
+					"name": "Delete",
+					"request": {
+						"method": "DELETE",
+						"header": [],
+						"url": {
+							"raw": "https://localhost:7167/api/categories/9",
+							"protocol": "https",
+							"host": [
+								"localhost"
+							],
+							"port": "7167",
+							"path": [
+								"api",
+								"categories",
+								"9"
+							]
+						}
+					},
+					"response": []
+				}
+			]
+		},
+		{
+			"name": "Product",
+			"item": [
+				{
+					"name": "List",
+					"request": {
+						"method": "GET",
+						"header": [],
+						"url": {
+							"raw": "https://localhost:7167/api/products?PageNumber=1&PageSize=10",
+							"protocol": "https",
+							"host": [
+								"localhost"
+							],
+							"port": "7167",
+							"path": [
+								"api",
+								"products"
+							],
+							"query": [
+								{
+									"key": "PageNumber",
+									"value": "1"
+								},
+								{
+									"key": "PageSize",
+									"value": "10"
+								}
+							]
+						}
+					},
+					"response": []
+				},
+				{
+					"name": "Get",
+					"request": {
+						"method": "GET",
+						"header": [],
+						"url": {
+							"raw": "https://localhost:7167/api/products/3",
+							"protocol": "https",
+							"host": [
+								"localhost"
+							],
+							"port": "7167",
+							"path": [
+								"api",
+								"products",
+								"3"
+							]
+						}
+					},
+					"response": []
+				},
+				{
+					"name": "Create",
+					"request": {
+						"method": "POST",
+						"header": [],
+						"body": {
+							"mode": "raw",
+							"raw": "{\r\n    \"name\":\"43\",\r\n    \"image\":\"\",\r\n    \"description\":\"Error\",\r\n    \"categoryId\": 4,\r\n    \"price\": 200,\r\n    \"amount\": 1\r\n}",
+							"options": {
+								"raw": {
+									"language": "json"
+								}
+							}
+						},
+						"url": {
+							"raw": "https://localhost:7167/api/products",
+							"protocol": "https",
+							"host": [
+								"localhost"
+							],
+							"port": "7167",
+							"path": [
+								"api",
+								"products"
+							]
+						}
+					},
+					"response": []
+				},
+				{
+					"name": "Update",
+					"request": {
+						"method": "PUT",
+						"header": [],
+						"body": {
+							"mode": "raw",
+							"raw": "{\r\n    \"id\": 3,\r\n    \"name\":\"Samsumg 43 pulgadas\",\r\n    \"image\":\"https://i.ytimg.com/vi/tCS2m0UDILc/maxresdefault.jpg\",\r\n    \"description\":\"Smart TV Tizen\",\r\n    \"categoryId\": 4,\r\n    \"price\": 2500,\r\n    \"amount\": 2\r\n}",
+							"options": {
+								"raw": {
+									"language": "json"
+								}
+							}
+						},
+						"url": {
+							"raw": "https://localhost:7167/api/products/3",
+							"protocol": "https",
+							"host": [
+								"localhost"
+							],
+							"port": "7167",
+							"path": [
+								"api",
+								"products",
+								"3"
+							]
+						}
+					},
+					"response": []
+				},
+				{
+					"name": "Delete",
+					"request": {
+						"method": "DELETE",
+						"header": [],
+						"url": {
+							"raw": "https://localhost:7167/api/products/5",
+							"protocol": "https",
+							"host": [
+								"localhost"
+							],
+							"port": "7167",
+							"path": [
+								"api",
+								"products",
+								"5"
+							]
+						}
+					},
+					"response": []
+				}
+			]
+		}
+	]
+}
